AangModel Rig Updated Further
Progress:

Added the L_leg and R_leg IKFK ctrls and foot controls as well but not constrained to geo at this point. Cog control was deleted at some point so I needed to remake the cog control and add it's constrains

Issues:

When the pelvis is moved, the legs will move in inverse directions of each other but move independently fine and in the correct orientation from what I can see in the inspector.

When moving the cog, there is a delay in movement between the legs and the right arm however, the left arm is stationary.